requirements.txt
asgiref==3.2.10
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
django-storages==1.9.1
django==3.0.8
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9; python_version != '3.4'
